Дорабатывает модуляцию. Выносит данные в отдельный файл
